Refactor: Rename compactdecimal.rs to formatter.rs and move the options to options.rs
