+++ components/experimental/src/compactdecimal/mod.rs
@@ -19,10 +19,11 @@
 )]
 #![warn(missing_docs)]
 
-mod compactdecimal;
 mod error;
 mod format;
+mod formatter;
+mod options;
 pub mod provider;
 
-pub use compactdecimal::CompactDecimalFormatter;
 pub use error::ExponentError;
+pub use formatter::CompactDecimalFormatter;
